miniRT_wip: Divided program, added Makefile.

+++ miniRT_wip/math/math.c
@@ -1,22 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   calc.c                                             :+:      :+:    :+:   */
+/*   math.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: zhedlund <zhedlund@student.42berlin.de>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/15 17:08:04 by zhedlund          #+#    #+#             */
-/*   Updated: 2024/04/15 17:11:40 by zhedlund         ###   ########.fr       */
+/*   Updated: 2024/04/16 14:46:36 by zhedlund         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "minirt.h"
+#include "../minirt.h"
 
 /*	calculates the dot product between two 3D vectors u and v.
 	repr cosine of the angle between them multiplied by the magnitudes of the vectors. 
 	u⋅v = ux⋅vx + uy⋅vy + uz⋅vz
 */
-double dot(const t_vec *u, const t_vec *v) 
+float dot(const t_vec *u, const t_vec *v) 
 {
     return ((u->x * v->x) + (u->y * v->y) + (u->z * v->z));
 }
@@ -24,7 +24,7 @@ double dot(const t_vec *u, const t_vec *v)
 /* calculates intersection point of a ray and a sphere.
    Returns a t_vec struct representing the intersection point.
 */
-t_vec intersect_point(const t_ray *r, double t)
+t_vec intersect_point(const t_ray *r, float t)
 {
     return ((t_vec){r->origin.x + (t * r->dir.x),
                    r->origin.y + (t * r->dir.y),
